feature/23/consolidate-db-queries
* Refactor, moving db queries to their own module
* fixes #23
