+++ app/pages/dashboard/add.vue
@@ -29,7 +29,7 @@ const onSubmit = handleSubmit(async (values) => {
     if (error.data?.data) {
       setErrors(error.data?.data);
     }
-    submitError.value = error.statusMessage || 'An unknown error occurred.';
+    submitError.value = error.data?.statusMessage || error.statusMessage || 'An unknown error occurred.';
   }
   finally {
     loading.value = false;
